# openapi.json
{"openapi": "3.1.0", "info": {"title": "AI SOC Firewall", "version": "1.1.0"}, "paths": {"/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/health": {"get": {"tags": ["Health"], "summary": "Health", "operationId": "health_api_v1_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/ready": {"get": {"tags": ["Health"], "summary": "Ready", "operationId": "ready_api_v1_ready_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/metrics": {"get": {"tags": ["Observability"], "summary": "Metrics", "operationId": "metrics_api_v1_metrics_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/auth/login": {"post": {"tags": ["Auth"], "summary": "Login", "operationId": "login_api_v1_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/me": {"get": {"tags": ["Auth"], "summary": "Me", "operationId": "me_api_v1_auth_me_get", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserProfile"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/attacks": {"get": {"tags": ["Attacks"], "summary": "List Attacks", "operationId": "list_attacks_api_v1_attacks_get", "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 1000, "minimum": 1, "default": 100, "title": "Limit"}}, {"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Attacks"], "summary": "Create Attack", "operationId": "create_attack_api_v1_attacks_post", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AttackLogIn"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/firewall/block": {"post": {"tags": ["Firewall"], "summary": "Block Ip", "operationId": "block_ip_api_v1_firewall_block_post", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FirewallActionRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FirewallActionResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/firewall/unblock": {"post": {"tags": ["Firewall"], "summary": "Unblock Ip", "operationId": "unblock_ip_api_v1_firewall_unblock_post", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FirewallActionRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FirewallActionResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/threat-intel/check": {"get": {"tags": ["Threat Intelligence"], "summary": "Check Ip Get", "operationId": "check_ip_get_api_v1_threat_intel_check_get", "parameters": [{"name": "ip", "in": "query", "required": true, "schema": {"type": "string", "title": "Ip"}}, {"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ThreatIntelResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Threat Intelligence"], "summary": "Check Ip Post", "operationId": "check_ip_post_api_v1_threat_intel_check_post", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ThreatIntelRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ThreatIntelResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/reports/generate": {"post": {"tags": ["Reports"], "summary": "Generate", "operationId": "generate_api_v1_reports_generate_post", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReportRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReportResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/hunting/search": {"post": {"tags": ["Threat Hunting"], "summary": "Search Hunts", "operationId": "search_hunts_api_v1_hunting_search_post", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ThreatHuntRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ThreatHuntResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/db/firewall-rules": {"get": {"tags": ["DB Admin"], "summary": "List Firewall Rules", "operationId": "list_firewall_rules_api_v1_db_firewall_rules_get", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/ingestion/file": {"post": {"tags": ["Log Ingestion"], "summary": "Ingest From File", "operationId": "ingest_from_file_api_v1_ingestion_file_post", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IngestFileRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/ml/predict": {"post": {"tags": ["ML"], "summary": "Predict", "operationId": "predict_api_v1_ml_predict_post", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MLPredictRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/geo/lookup": {"get": {"tags": ["Geo"], "summary": "Geo Lookup", "operationId": "geo_lookup_api_v1_geo_lookup_get", "parameters": [{"name": "ip", "in": "query", "required": true, "schema": {"type": "string", "title": "Ip"}}, {"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/siem/export": {"get": {"tags": ["SIEM"], "summary": "Export Siem", "operationId": "export_siem_api_v1_siem_export_get", "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 5000, "minimum": 1, "default": 500, "title": "Limit"}}, {"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"AttackLogIn": {"properties": {"source_ip": {"type": "string", "title": "Source Ip"}, "destination_ip": {"type": "string", "title": "Destination Ip"}, "attack_type": {"type": "string", "maxLength": 100, "minLength": 2, "title": "Attack Type"}, "severity": {"type": "string", "title": "Severity"}, "timestamp": {"type": "string", "format": "date-time", "title": "Timestamp"}, "raw_message": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Raw Message", "default": ""}, "status_code": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Status Code", "default": 0}, "payload": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Payload", "default": ""}}, "type": "object", "required": ["source_ip", "destination_ip", "attack_type", "severity", "timestamp"], "title": "AttackLogIn"}, "FirewallActionRequest": {"properties": {"ip_address": {"type": "string", "title": "Ip Address"}, "reason": {"type": "string", "maxLength": 200, "minLength": 2, "title": "Reason", "default": "SOC analyst action"}}, "type": "object", "required": ["ip_address"], "title": "FirewallActionRequest"}, "FirewallActionResponse": {"properties": {"success": {"type": "boolean", "title": "Success"}, "message": {"type": "string", "title": "Message"}, "ip_address": {"type": "string", "title": "Ip Address"}, "action": {"type": "string", "title": "Action"}}, "type": "object", "required": ["success", "message", "ip_address", "action"], "title": "FirewallActionResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "IngestFileRequest": {"properties": {"file_path": {"type": "string", "title": "File Path"}}, "type": "object", "required": ["file_path"], "title": "IngestFileRequest"}, "LoginRequest": {"properties": {"email": {"type": "string", "maxLength": 120, "minLength": 3, "title": "Email"}, "password": {"type": "string", "minLength": 1, "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "LoginRequest"}, "MLPredictRequest": {"properties": {"source_port": {"type": "integer", "title": "Source Port"}, "dest_port": {"type": "integer", "title": "Dest Port"}, "bytes_sent": {"type": "integer", "title": "Bytes Sent"}, "bytes_received": {"type": "integer", "title": "Bytes Received"}, "failed_logins": {"type": "integer", "title": "Failed Logins"}, "request_rate": {"type": "integer", "title": "Request Rate"}, "is_internal_src": {"type": "integer", "title": "Is Internal Src"}, "proto": {"type": "string", "title": "Proto"}, "severity_num": {"type": "integer", "title": "Severity Num"}}, "type": "object", "required": ["source_port", "dest_port", "bytes_sent", "bytes_received", "failed_logins", "request_rate", "is_internal_src", "proto", "severity_num"], "title": "MLPredictRequest"}, "ReportRequest": {"properties": {"incident_id": {"type": "string", "title": "Incident Id"}}, "type": "object", "required": ["incident_id"], "title": "ReportRequest"}, "ReportResponse": {"properties": {"report_name": {"type": "string", "title": "Report Name"}, "report_path": {"type": "string", "title": "Report Path"}, "generated_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Generated At"}}, "type": "object", "required": ["report_name", "report_path"], "title": "ReportResponse"}, "ThreatHuntRequest": {"properties": {"source_ip": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Source Ip"}, "attack_type": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Attack Type"}, "severity": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Severity"}}, "type": "object", "title": "ThreatHuntRequest"}, "ThreatHuntResponse": {"properties": {"total": {"type": "integer", "title": "Total"}, "results": {"items": {"type": "object"}, "type": "array", "title": "Results"}}, "type": "object", "required": ["total", "results"], "title": "ThreatHuntResponse"}, "ThreatIntelRequest": {"properties": {"ip": {"type": "string", "title": "Ip"}}, "type": "object", "required": ["ip"], "title": "ThreatIntelRequest"}, "ThreatIntelResponse": {"properties": {"ip": {"type": "string", "title": "Ip"}, "reputation_score": {"type": "integer", "title": "Reputation Score"}, "malicious": {"type": "boolean", "title": "Malicious"}, "country": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Country"}, "isp": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Isp"}, "source": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Source", "default": "mock"}}, "type": "object", "required": ["ip", "reputation_score", "malicious"], "title": "ThreatIntelResponse"}, "TokenResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}, "role": {"type": "string", "title": "Role"}}, "type": "object", "required": ["access_token", "role"], "title": "TokenResponse"}, "UserProfile": {"properties": {"id": {"type": "string", "title": "Id"}, "email": {"type": "string", "title": "Email"}, "full_name": {"type": "string", "title": "Full Name"}, "role": {"type": "string", "title": "Role"}}, "type": "object", "required": ["id", "email", "full_name", "role"], "title": "UserProfile"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}